refactor(dt): extract resolveHeaders and getServerTraceparent helpers
Code-review cleanup — relocate pure helpers out of large/component files:
- Move resolveHeaders from provider.tsx to utils.ts (alongside the other
  pure header/SDK helpers); move its unit tests to utils.test.ts. Fix the
  stale comment (it resolves at construction, not per-request).
- Move getServerTraceparent (server-only traceparent resolver with the
  WEBPACK_TARGET guard + lazy require) from the template _app-config into
  app/utils/tracing-utils.js, keeping the config component clean.

No behavior change. provider.tsx left otherwise cohesive (a structural
split into single-use hooks would add encapsulation overhead, not clarity).
