+++ packages/commerce-sdk-react/src/utils.test.ts
@@ -224,4 +224,37 @@ describe('Utils', () => {
             expect(configWithCustomProps.transformer).toHaveBeenCalled()
         })
     })
+
+    describe('resolveHeaders', () => {
+        test('string values pass through unchanged', () => {
+            expect(utils.resolveHeaders({'correlation-id': 'static-value'})).toEqual({
+                'correlation-id': 'static-value'
+            })
+        })
+
+        test('callable values are invoked and their return value is used', () => {
+            const getTraceparent = jest.fn(() => '00-trace-span-01')
+            const result = utils.resolveHeaders({traceparent: getTraceparent})
+            expect(getTraceparent).toHaveBeenCalledTimes(1)
+            expect(result).toEqual({traceparent: '00-trace-span-01'})
+        })
+
+        test('callable returning undefined is excluded from the result', () => {
+            const result = utils.resolveHeaders({
+                'correlation-id': 'keep-me',
+                traceparent: () => undefined
+            })
+            expect(result).toEqual({'correlation-id': 'keep-me'})
+            expect(result).not.toHaveProperty('traceparent')
+        })
+
+        test('callable returning empty string is excluded from the result', () => {
+            const result = utils.resolveHeaders({
+                'correlation-id': 'keep-me',
+                traceparent: () => ''
+            })
+            expect(result).toEqual({'correlation-id': 'keep-me'})
+            expect(result).not.toHaveProperty('traceparent')
+        })
+    })
 })
